NAS-126998: fixed tests and moved chained-component-ref class
